refactor: extracting share responsibility and creating useTerritoryData
